requirements.txt
addict==2.4.0
joblib==1.4.2
numpy==1.26.4
omegaconf==2.3.0
pandas==2.2.3
torch==2.5.0
pytorch_forecasting==1.1.1
pytorch_lightning==2.4.0
pytorch_tsmixer==0.2.0
scikit_learn==1.5.2
scipy==1.14.1
tft_torch==0.0.6
tqdm==4.66.5
tensorboard==2.18.0
